Refactor code into a NumberFormatter class.
NumberFormatter has a factory method to return an instance that
formats for a specific language. The English formatter is created
by default. In theory, formatters for other languages can be plugged
in by registering them at runtime, passing a map of translated strings.
This has yet to be validated, and there are almost certainlly kinks
to be worked out for non-English formatters - consider this a
starting point.