Implemented scoring for all poker hands.
I changed the hand-scoring mechanism a bit - each matcher function
just returns the card-value (eg: high-card), and the relative value
of different hands (eg: flush vs. straight) is set externally, in
a "game", which is just a list of match-rules and their values.

Test coverage is not as much as I would like yet - I'll add some
more when I find the time.